Move utility zip functions into plugin/contrib/mesos/pkg/archive module
